add period validation

+++ src/config.rs
@@ -12,16 +12,16 @@ pub struct Config {
 struct Cli {
     #[arg(short, long)]
     username: String,
-    #[arg(short, long, value_parser = clap::value_parser!(u32).range(1..=10))]
-    limit: Option<u32>,
-    #[arg(short, long)]
+    #[arg(short, long, value_parser = clap::value_parser!(u8).range(1..=10))]
+    limit: Option<u8>,
+    #[arg(short, long, value_parser = ["7day", "1month", "6month", "12month", "overall"])]
     period: Option<String>,
 }
 
 impl Config {
     pub fn build() -> Config {
         let cli = Cli::parse();
-        let default_limit: u32 = 5;
+        let default_limit: u8 = 5;
         let default_period = "7day".to_string();
 
         let final_limit = cli.limit.unwrap_or(default_limit);
